bowtie2-build and bowtie2-align error fixed
"bowtie2-build" was being called with the argument "--threads $CPU", which crashes the program because "bowtie2-build" doesn't accept such argument (it can't be multithreaded).

"bowtie2" does support multithreading, but the argument is not "--threads", but "-p".

It's my first time doing this, hope you don't mind. Have a nice day.